FINERACT-2169: Accrual account api refactor;
